requirements.txt
scrapy_redis==0.6.8
selenium==3.6.0
pymongo==3.5.1
Scrapy==1.5.1
redis==2.10.6
Pillow==5.2.0
lxml==4.2.5
scrapy-redis-bloomfilter==0.7.0
python-dateutil==2.8.1